- Several functions edited and fingerprint routine added.
A new user was added for the beginning of adding a feature to allow someone without an account to send an encrypted message to a user with an account.  This could work like a form for support or for sending other sensitive information.  The main feature is that anyone can send someone with an account a message without having to create one themselves.